Consolidate validate_unique_values into utils/validation
