+++ baybe/parameters/numerical.py
@@ -13,10 +13,9 @@ from typing_extensions import override
 
 from baybe.exceptions import NumericalUnderflowError
 from baybe.parameters.base import ContinuousParameter, DiscreteParameter
-from baybe.parameters.validation import validate_unique_values
 from baybe.settings import active_settings
 from baybe.utils.interval import InfiniteIntervalError, Interval
-from baybe.utils.validation import validate_is_finite
+from baybe.utils.validation import validate_is_finite, validate_unique_values
 
 
 @define(frozen=True, slots=False)
