+++ baybe/utils/validation.py
@@ -3,7 +3,8 @@
 from __future__ import annotations
 
 import math
-from collections.abc import Callable, Iterable, Sequence
+from collections import Counter
+from collections.abc import Callable, Collection, Iterable, Sequence
 from typing import TYPE_CHECKING, Any
 
 import numpy as np
@@ -276,4 +277,20 @@ def validate_is_finite(  # noqa: DOC101, DOC103
         raise ValueError(
             f"Cannot assign the following values containing infinity/nan to "
             f"'{attribute.alias}': {value}."
+        )
+
+
+def validate_unique_values(  # noqa: DOC101, DOC103
+    _: Any, attribute: Attribute, value: Collection[str]
+) -> None:
+    """Validate that there are no duplicates in ``value``.
+
+    Raises:
+        ValueError: If there are duplicates in ``value``.
+    """
+    duplicates = [item for item, count in Counter(value).items() if count > 1]
+    if duplicates:
+        raise ValueError(
+            f"Entries appearing multiple times: {duplicates}. "
+            f"All entries of '{attribute.alias}' must be unique."
         )
